+++ src/gx/main.go
@@ -19,13 +19,17 @@ package main
 import (
 	"com.tuntun.rangers/node/src/gx/cli"
 	"fmt"
+	"os"
+	"os/signal"
 	"runtime"
 	"runtime/debug"
+	"syscall"
+	"time"
 )
 
 func main() {
 	initSysParam()
-	SetupStackTrap()
+	setupStackTrap()
 	gx := cli.NewGX()
 	gx.Run()
 
@@ -37,4 +41,42 @@ func initSysParam() {
 	debug.SetMaxStack(1 * 1000 * 1000 * 1000)
 
 	fmt.Printf("Setting gc %s, max memory %s, maxproc %s\n", "50", "1g", "8")
+}
+
+const (
+	timeFormat = "2006-01-02 15:04:05"
+)
+
+var (
+	stdFile = "./stack.log"
+)
+
+func setupStackTrap(args ...string) {
+	if len(args) > 0 {
+		stdFile = args[0]
+	}
+
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, syscall.SIGUSR1)
+	go func() {
+		for range c {
+			dumpStacks()
+		}
+	}()
+}
+
+func dumpStacks() {
+	buf := make([]byte, 1638400)
+	buf = buf[:runtime.Stack(buf, true)]
+	writeStack(buf)
+}
+
+func writeStack(buf []byte) {
+	fd, _ := os.OpenFile(stdFile, os.O_CREATE|os.O_RDWR|os.O_APPEND, 0666)
+
+	now := time.Now().Format(timeFormat)
+	fd.WriteString("\n\n\n\n\n")
+	fd.WriteString(now + " stdout:" + "\n\n")
+	fd.Write(buf)
+	fd.Close()
 }
